+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/exception/NotFoundHandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.techcourse.exception;
+package webmvc.org.springframework.web.servlet.mvc.exception;
 
 public class NotFoundHandlerAdapter extends IllegalArgumentException {
